--- UniMarket-Swift/UniMarket-Swift/components/EmptyStateView.swift
@@ -1,7 +1,0 @@
-//
-//  EmptyStateView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
