--- UniMarket-Swift/UniMarket-Swift/components/PrimaryButton.swift
@@ -1,7 +1,0 @@
-//
-//  PrimaryButton.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
