--- UniMarket-Swift/UniMarket-Swift/components/ProductCardView.swift
@@ -1,7 +1,0 @@
-//
-//  ProductCardView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
