--- UniMarket-Swift/UniMarket-Swift/models/Order.swift
@@ -1,7 +1,0 @@
-//
-//  Order.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
