--- UniMarket-Swift/UniMarket-Swift/repositories/LocalCache.swift
@@ -1,7 +1,0 @@
-//
-//  LocalCache.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
