--- UniMarket-Swift/UniMarket-Swift/repositories/SecureStore.swift
@@ -1,7 +1,0 @@
-//
-//  SecureStore.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
