--- UniMarket-Swift/UniMarket-Swift/services/AuthService.swift
@@ -1,37 +1,0 @@
-import Foundation
-import FirebaseAuth
-import FirebaseFirestore
-import Combine
-
-class AuthService: ObservableObject {
-    @Published var userSession: FirebaseAuth.User?
-    @Published var currentUser: User?
-    
-    static let shared = AuthService()
-    
-    init() {
-        self.userSession = Auth.auth().currentUser
-        Task { await fetchUser() }
-    }
-    
-    @MainActor
-    func fetchUser() async {
-        guard let uid = Auth.auth().currentUser?.uid else { return }
-        
-        do {
-            let snapshot = try await Firestore.firestore().collection("users").document(uid).getDocument()
-            self.currentUser = try snapshot.data(as: User.self)
-        } catch {
-            print("DEBUG: Failed to fetch user with error \(error.localizedDescription)")
-        }
-    }
-    
-    @MainActor
-    func updateProfileImage(withImageUrl imageUrl: String) async throws {
-        guard let uid = Auth.auth().currentUser?.uid else { return }
-        try await Firestore.firestore().collection("users").document(uid).updateData([
-            "profilePic": imageUrl
-        ])
-        self.currentUser?.profilePic = imageUrl
-    }
-}
