--- UniMarket-Swift/UniMarket-Swift/utils/Constants.swift
@@ -1,7 +1,0 @@
-//
-//  Constants.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
