--- UniMarket-Swift/UniMarket-Swift/viewmodels/AuthViewModel.swift
@@ -1,7 +1,0 @@
-//
-//  AuthViewModel.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
