--- UniMarket-Swift/UniMarket-Swift/views/CartView.swift
@@ -1,7 +1,0 @@
-//
-//  CartView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
