--- UniMarket-Swift/UniMarket-Swift/views/CheckoutView.swift
@@ -1,7 +1,0 @@
-//
-//  CheckoutView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
