--- UniMarket-Swift/UniMarket-Swift/views/MyPurchasesView.swift
@@ -1,7 +1,0 @@
-//
-//  MyPurchasesView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
