--- UniMarket-Swift/UniMarket-Swift/views/ProfileEditView.swift
@@ -1,7 +1,0 @@
-//
-//  ProfileEditView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
