--- UniMarket-Swift/UniMarket-Swift/views/SellView.swift
@@ -1,7 +1,0 @@
-//
-//  SellView.swift
-//  UniMarket-Swift
-//
-//  Created by Mariana Pineda on 1/03/26.
-//
-
